requirements/test.txt
-r dev.txt
pytest==9.0.2
pytest-django==4.12.0
